fix: add some eslint rules
